Refactor generate metadata into multiple files
