+++ metadata/__init__.py
@@ -1,0 +1,1 @@
+"""RH Community Plugins metadata scripts."""
